Issue #251 - mg_control does not display readout stats/states
*  Added new ReadoutStatus.py module to the installation of modules.
   aded a model and view for stats.
*  Removed the readout state model from maingui moving it to
ReadoutStatus.py
*  Wrote the ReadoutStatusView in ReadoutStatus.py
