+++ main/utilities/managercontrol/maingui.py
@@ -18,7 +18,7 @@ from PyQt5.QtWidgets import (
     QTextEdit
     
 )
-from PyQt5.QtGui import ( QStandardItemModel, QStandardItem, QIntValidator)
+from PyQt5.QtGui import (QIntValidator)
 
 from PyQt5.QtCore import pyqtSignal, Qt
 
@@ -440,97 +440,7 @@ class RunControlWidget(QWidget):
         else:
             self._log.setDisabled(False)
 
-class ReadoutStatusModel(QStandardItemModel) :
-    '''
-        This model is intended to be used with a table for the
-        states/statistics of the readout programs.  It maintains the following
-        columns:
-        * name - the name of the readout program.
-        * state - The state of the readout program which can be one of:
-            - Discon       - the program is not reachable via ReST.
-            - Idle         - the program is connected but the run is not active.
-            - Active       - The program is connected and a run is Active.
-        * cum trg - cummulative triggers.
-        * cum atrg - Cummulative accetped triggers
-        * cum bytes - Cummulative data event payload bytes.
-        * run trg   - Triggers since last begin run.
-        * run atrg  - Triggers accepted since last begin run.
-        * run bytes - bytes of event data payload since last begin run.
-    '''
-    def __init__(self):
-        super.__init__(self)
-        
-        # Define the headings.  We assume the 
-        # table has this header turned on.
-        
-        self.setHorizontalHeaderLabels((
-            "Name", "State", 
-            "Cum trg", "cum atrg", "cum bytes",
-            "run trg", "run atrg", "run bytes"
-        ))
-        
-    #   Return a dict indexed by program name whose values are the line of the table.
-    def _enumeratePrograms(self):
-        rows = self.rowCount()
-        result = dict()
-        
-        for row in range(0, rows):
-            name = self.item(row, 0).text()
-            result[name] = row
-        
-        return result
-        
-        
-    #  Given a list of dicts with a 'name' key, returns the list of name values
-    #  in that list
-    
-    @staticmethod
-    def _present(info):
-        result = []
-        for item in info:
-            result.append(item['name'])
-        return result
-    
-    def update(self, info):
-        '''
-            We update the table from an iterable of dicts where each dict has been returned
-            from getStatistics and augmented with a 'name' and 'state' key that has the program name.
-            
-            If a name exists in the model, we just update the rest of the line.
-            If a name does not yet exist in the model, it is added.  If the model has a name
-            but it's not in the dict, the name is marekd with the 'Discon' state.
-        '''
-        
-        #  Enumerate names in the model in a dict indexed by name with value the row the name is in:
-        
-        existingPrograms  = self._enumeratePrograms()
-        names = existingPrograms.keys()
-        for item in info :
-            if item['name'] not in names:
-                existingPrograms[item['name']] = self.rowCount()
-                self.appendRow(QStandardItem(item['name']))   # Make a new row.
-                
-            # Update the row rest of the row in place:
-            row = existingPrograms[item['name']]
-            self.setItem(row, 1, QStandardItem(item['state']))
-            
-            self.setItem(row, 2, QStandardItem(item['cumulative']['triggers']))
-            self.setItem(row, 3, QStandardItem(item['cumulative']['acceptedTriggers']))
-            self.setItem(row, 4, QStandardItem(item['cumulative']['bytes']))
-            
-            self.setItem(row, 5, QStandardItem(item['perRun']['triggers']))    
-            self.setItem(row, 6, QStandardItem(item['perRun']['acceptedTriggers']))
-            self.setItem(row, 7, QStandardItem(item['perRun']['bytes']))
-        
-        # Now Set the state of rows that are not in info list
-        
-        connectedNames = ReadoutStatusModel._present(info)
-        for name in names:
-            if name not in connectedNames:
-                # Not connected.
-                self.setItem(existingPrograms[name], 1, QStandardItem('Discon'))
-        
-        
+
 class RunControls(QWidget):
     """Container widget for the RunInfoWidget and RunControlWidgets.  These are the top part of the
     main GUI
@@ -959,9 +869,9 @@ if __name__ == "__main__":
     import sys
     app = QApplication(sys.argv)
     main_window = QMainWindow()
-    test_widget = MainGui()
+    main_widget = MainGui()
     
-    main_window.setCentralWidget(test_widget)
+    main_window.setCentralWidget(main_widget)
     
 
     main_window.show()
